+++ diive/flux/__init__.py
@@ -1,4 +1,9 @@
-from diive.pkgs.flux.fluxprocessingchain.fluxprocessingchain import FluxProcessingChain
+from diive.pkgs.flux.fluxprocessingchain import (
+    FluxConfig,
+    FluxLevelData,
+    FluxProcessingChain,
+    run_flux_chain,
+)
 from diive.pkgs.flux.hires.fluxdetectionlimit import FluxDetectionLimit
 from diive.pkgs.flux.hires.lag import MaxCovariance
 from diive.pkgs.flux.hires.lag_pwb import PreWhiteningBootstrap